Make winning visualization interaction a bit smoother.
Sets a class attribute to the value of whatever visualization color is
currently active, for use in CSS where needed.

Adds styles for various states for specific words in a Prism on the winning
visualization.

Also helps fix a CSS issue in Safari where the pointer cursor was being
displayed on highlight pages instead of specific marker images.