Changed the way the word frequencies are sorted for winning facet viz.
For some reason, Safari wasn't reversing the list, based on the
comparison function. This fixes that.

Rather than using a user-defined comparison function, I'm using a
Schwartzian transformation and two calls to native methods (`sort` and
`reverse`), which I expect is still faster than calling into a
user-defined function for each comparison that the sort makes. Of
course, for a list of this size (3), it's probably a toss-up.

Also, I built the word frequencies list using a comprehension.

Fixes #280.